+++ test/absinthe/execution/subscription_test.exs
@@ -3,40 +3,6 @@ defmodule Absinthe.Execution.SubscriptionTest do
 
   import ExUnit.CaptureLog
 
-  defmodule PubSub do
-    @behaviour Absinthe.Subscription.Pubsub
-
-    def start_link() do
-      Registry.start_link(keys: :unique, name: __MODULE__)
-    end
-
-    def node_name() do
-      node()
-    end
-
-    def subscribe(topic) do
-      Registry.register(__MODULE__, topic, [])
-      :ok
-    end
-
-    def publish_subscription(topic, data) do
-      message = %{
-        topic: topic,
-        event: "subscription:data",
-        result: data
-      }
-
-      Registry.dispatch(__MODULE__, topic, fn entries ->
-        for {pid, _} <- entries, do: send(pid, {:broadcast, message})
-      end)
-    end
-
-    def publish_mutation(_proxy_topic, _mutation_result, _subscribed_fields) do
-      # this pubsub is local and doesn't support clusters
-      :ok
-    end
-  end
-
   defmodule Schema do
     use Absinthe.Schema
 
@@ -154,8 +120,8 @@ defmodule Absinthe.Execution.SubscriptionTest do
   end
 
   setup_all do
-    {:ok, _} = PubSub.start_link()
-    {:ok, _} = Absinthe.Subscription.start_link(PubSub)
+    {:ok, _} = start_supervised(PubSub)
+    {:ok, _} = start_supervised({Absinthe.Subscription, [PubSub]})
     :ok
   end
